LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/worker-shared-blocked.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="/security/contentSecurityPolicy/resources/child-src-test.js"></script>
    <meta http-equiv="Content-Security-Policy" content="child-src 'none';">
</head>
<body>
    <script>
        description("SharedWorkers should be governed by 'child-src'.");
        injectSharedWorker("/security/contentSecurityPolicy/resources/alert-fail.js", EXPECT_BLOCK);
    </script>
</body>
</html>
