LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/worker-shared-blocked-expected.txt

CONSOLE MESSAGE: Refused to create a child context containing 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.js' because it violates the following Content Security Policy directive: "child-src 'none'".

SharedWorkers should be governed by 'child-src'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS var w = new SharedWorker('/security/contentSecurityPolicy/resources/alert-fail.js'); threw exception SecurityError: Failed to construct 'SharedWorker': Access to the script at 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.js' is denied by the document's Content Security Policy..
PASS successfullyParsed is true

TEST COMPLETE
